Refactor manager datacenter
