+++ manager.go
@@ -42,12 +42,6 @@ type Service struct {
 	Endpoint   string    `json:"endpoint"`
 }
 
-// Datacenter ...
-type Datacenter struct {
-	ID   int    `json:"id"`
-	Name string `json:"name"`
-}
-
 // Session ...
 type Session struct {
 	UserID  int    `json:"id"`
@@ -122,19 +116,6 @@ func (m *Manager) getSession(token string) (session Session, err error) {
 		return session, err
 	}
 	return session, nil
-}
-
-// ********************* Create *******************
-
-// CreateDatacenter ...
-func (m *Manager) CreateDatacenter(token string, name string, user string, password string, url string, network string, vseURL string) (string, error) {
-	payload := []byte(`{"name": "` + name + `", "type": "vcloud", "region": "LON-001", "username":"` + user + `", "password":"` + password + `", "external_network":"` + network + `", "vcloud_url":"` + url + `", "vse_url":"` + vseURL + `"}`)
-	body, _, err := m.doRequest("/api/datacenters/", "POST", payload, token, "")
-	if err != nil {
-		return body, err
-	}
-	color.Green("SUCCESS: Datacenter " + name + " created")
-	return body, err
 }
 
 // ********************* Get *******************
@@ -255,19 +236,6 @@ func (m *Manager) ServiceBuildStatus(token string, serviceName string, serviceID
 }
 
 // ********************* List *********************
-
-// ListDatacenters ...
-func (m *Manager) ListDatacenters(token string) (datacenters []Datacenter, err error) {
-	body, _, err := m.doRequest("/api/datacenters/", "GET", []byte(""), token, "")
-	if err != nil {
-		return nil, err
-	}
-	err = json.Unmarshal([]byte(body), &datacenters)
-	if err != nil {
-		return nil, err
-	}
-	return datacenters, err
-}
 
 // ListServices ...
 func (m *Manager) ListServices(token string) (services []Service, err error) {
